pyproject.toml
[project]
name = "Xlsx"
version = "0.1.0"
description = "NoneBot2 Excel插件 - 支持Excel文件导入导出、数据管理和自动代肝记录"
readme = "README.md"
requires-python = ">=3.9, <4.0"
dependencies = [
    "nonebot2[fastapi]>=2.3.0",
    "nonebot-adapter-onebot>=2.4.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.0.0"
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = ["./plugins"]
builtin_plugins = []
